fix: UpdateLocally in MailboxManager+Local
